requirements.txt
accelerate==1.7.0
acres==0.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aistudio-sdk==0.3.6
alibabacloud-credentials==1.0.2
alibabacloud-credentials-api==1.0.0
alibabacloud-tea==0.4.3
alibabacloud_opensearch_util==0.0.4
alibabacloud_tea_util==0.3.13
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.9.0
async-timeout==4.0.3
asyncpg==0.30.0
attrs==25.3.0
backoff==2.2.1
bce-python-sdk==0.9.45
beautifulsoup4==4.13.4
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
celery==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorlog==6.9.0
ConfigArgParse==1.7
configobj==5.0.9
configparser==7.2.0
cryptography==45.0.4
dashscope==1.24.2
dataclasses-json==0.6.7
datasets==3.6.0
dill==0.3.8
distro==1.9.0
elastic-transport==8.17.1
elasticsearch==8.13.2
emoji==2.14.1
et_xmlfile==2.0.0
etelemetry==0.3.1
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
flower==2.0.1
frozenlist==1.6.0
fsspec==2025.3.0
future==1.0.0
gevent==25.4.2
geventhttpclient==2.3.3
greenlet==3.2.1
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.8
httplib2==0.30.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanize==4.12.3
idna==3.10
imagesize==1.4.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
isodate==0.7.2
itsdangerous==2.2.0
jieba==0.42.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jq==1.9.1
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.3
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.3.45
locust==2.34.0
looseversion==1.3.0
lxml==5.4.0
Markdown==3.8.2
MarkupSafe==3.0.2
marshmallow==3.26.1
minio==7.2.16
modelscope==1.29.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
multiprocess==0.70.16
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.2.1
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numpy==2.0.2
olefile==0.47
openai==1.73.0
opencv-contrib-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.10.18
packaging==24.2
paddleocr==3.2.0
paddlepaddle==3.1.1
paddlex==3.2.1
pandas==2.2.3
pathlib==1.0.1
pdfminer.six==20250506
pillow==11.2.1
prettytable==3.16.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.32.0
prov==2.1.1
psutil==7.0.0
psycopg2==2.9.10
puremagic==1.30
py-cpuinfo==9.0.0
pyahocorasick==2.2.0
pyarrow==20.0.0
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydot==4.0.1
PyMuPDF==1.24.8
PyMuPDFb==1.24.8
pyparsing==3.2.3
pypdf==5.6.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==26.4.0
RapidFuzz==3.13.0
rdflib==7.1.4
redis==6.0.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.0.2
shapely==2.0.7
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.46.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.6.0
tornado==6.5.1
tqdm==4.67.1
traits==7.0.2
transformers==4.51.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
unstructured==0.17.2
unstructured-client==0.37.1
urllib3==2.4.0
uvicorn==0.34.1
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
xlrd==2.0.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
